updated wrapper libraries
git-svn-id: svn+ssh://svn.code.sf.net/p/tisch/code/libtisch-2.0@359 28952012-f250-0410-9090-e835bad21ca5

+++ libs/wrappers/java/libtisch/BlobCount.java
@@ -35,7 +35,7 @@ public class BlobCount extends FeatureInt {
     super.delete();
   }
 
-  public BlobCount(int tf) {
+  public BlobCount(long tf) {
     this(libtischJNI.new_BlobCount__SWIG_0(tf), true);
   }
