+++ libs/wrappers/java/libtisch/BlobID.java
@@ -35,7 +35,7 @@ public class BlobID extends FeatureInt {
     super.delete();
   }
 
-  public BlobID(int tf) {
+  public BlobID(long tf) {
     this(libtischJNI.new_BlobID__SWIG_0(tf), true);
   }
