+++ libs/wrappers/java/libtisch/FeatureBase.java
@@ -50,7 +50,7 @@ public class FeatureBase {
     libtischJNI.FeatureBase_change_ownership(this, swigCPtr, true);
   }
 
-  public FeatureBase(int _tf) {
+  public FeatureBase(long _tf) {
     this(libtischJNI.new_FeatureBase__SWIG_0(_tf), true);
     libtischJNI.FeatureBase_director_connect(this, swigCPtr, swigCMemOwn, true);
   }
@@ -93,11 +93,11 @@ public class FeatureBase {
     return libtischJNI.FeatureBase_has_result_get(swigCPtr, this);
   }
 
-  protected void setTypeflags(int value) {
+  protected void setTypeflags(long value) {
     libtischJNI.FeatureBase_typeflags_set(swigCPtr, this, value);
   }
 
-  protected int getTypeflags() {
+  protected long getTypeflags() {
     return libtischJNI.FeatureBase_typeflags_get(swigCPtr, this);
   }
